Make libuwb_uci_jni_rust_tests available to the platform
libuwb_uci_jni_rust_tests was hitting a quirk in Soong where a module
that was not available for the platform but had no apexes that depended
on it was treated as a platform variant.  That quirk is being fixed as
part of preparing the apex info mutator for incremental analysis.
Explicitly mark libuwb_uci_jni_rust_tests as available to the platform.

Bug: 372543712
Test: builds
Change-Id: I4e60e591750136bc1e8b4c2972316871820f033e