Replace .toList() with .collect()
In android, .toList() was only introduced in API level 34, but this
module has a min_sdk_version of 33. Use .collect() instead, which
also allows us to collect directly to a set and avoid a copy.

This was found when updating android lint, the old version of the
NewApi linter missed this.

Bug: 383061307
Test: Presubmits
Change-Id: I80ea2ef142b57443994ea4d04ad88e49183eb15d